+++ primebool.c
@@ -11,7 +11,7 @@ bool isPrime(int n)
         return false;
 
     // Check for divisibility from 2 to sqrt(n)
-    for (int i = 2; i <= sqrt(n); i++) {
+    for (int i = 2; i <= (int)sqrt(n); i++) {
         if (n % i == 0)
             return false;
     }
